config: Organize src directory and .c files include/minishell.h - Fix norm errors. - 90-96: Reorganize prototypes.
src/util
- Move and organize utility functions source files into util directory.
- Split functions in 'ft_strtok.c' into dedicated new files.
