+++ src/util/ft_strcspn.c
