+++ src/util/ft_strncpy.c
